refactor: rename all FAQFooter entries to Footer
