+++ src/layouts/AuthLayout.tsx
@@ -1,7 +1,7 @@
 import { Suspense, FC } from "react";
 import { Outlet } from "react-router";
 
-import FAQFooter from "../components/FAQFooter/FAQFooter";
+import Footer from "../components/Footer/Footer";
 import Header from "../components/Header/Header";
 import Preloader from "../components/FallBackPreloader/FallBackPreloader";
 
@@ -12,7 +12,7 @@ const AuthLayout: FC = () => {
       <Suspense fallback={<Preloader />}>
         <Outlet />
       </Suspense>
-      <FAQFooter links={[]} />
+      <Footer links={[]} />
     </>
   );
 };
