+++ src/layouts/FAQLayout.tsx
@@ -1,7 +1,7 @@
 import { Suspense, FC } from "react";
 import { Outlet } from "react-router";
 
-import FAQFooter from "../components/FAQFooter/FAQFooter";
+import Footer from "../components/Footer/Footer";
 import Preloader from "../components/FallBackPreloader/FallBackPreloader";
 import Header from "../components/Header/Header";
 import { FOOTER_LINKS } from "../App";
@@ -13,7 +13,7 @@ const HomeLayout: FC = () => {
       <Suspense fallback={<Preloader />}>
         <Outlet />
       </Suspense>
-      <FAQFooter links={FOOTER_LINKS} />
+      <Footer links={FOOTER_LINKS} />
     </>
   );
 };
